Fix incomplete plan items from verification review
Session 2 verification found several items marked complete but not actually
implemented. This commit fixes them:

- 0B.4: Fix RRSIG timestamps to u32 in dnssec_impl.rs (RFC 4034 compliance)
- 2D.5: Transform cache now uses moka::Cache with byte-size based eviction
- 3G.5: UploadValidator fully wired into HTTP request pipeline
- 6D.3: Connection semaphore (10k default) added to HTTP server
- 6E.1: WASM instance pool stub created at src/plugin/instance_pool.rs
- 6H.1: Role comparisons updated to use is_global() for composite roles
- 5A.3: Serverless documentation added at docs/SERVERLESS.md

Also:
- Update AGENTS.md with Session Lessons Learned section
- Update plans/plan.md with verification session notes

Verification:
- cargo clippy --lib -- -D warnings: PASS
- cargo test --test integration_test: 78/78 PASS

+++ src/dns/server/dnssec_impl.rs
@@ -541,8 +541,8 @@ impl DnsServer {
         rrsig.push(key.algorithm.to_u8());
         rrsig.push(labels);
         rrsig.extend_from_slice(&record.ttl.to_be_bytes());
-        rrsig.extend_from_slice(&sig_expire.to_be_bytes());
-        rrsig.extend_from_slice(&sig_inception.to_be_bytes());
+        rrsig.extend_from_slice(&(sig_expire as u32).to_be_bytes());
+        rrsig.extend_from_slice(&(sig_inception as u32).to_be_bytes());
         rrsig.extend_from_slice(&key.key_tag.to_be_bytes());
 
         let signer = signer_name.trim_end_matches('.');
